分离 page 核心方法
